+++ program-test/tests/cpi.rs
@@ -4,13 +4,12 @@ use {
     solana_keypair::Keypair,
     solana_msg::msg,
     solana_program::{instruction::get_stack_height, program::invoke},
-    solana_program_entrypoint::{MAX_PERMITTED_DATA_INCREASE, ProgramResult},
+    solana_program_entrypoint::{MAX_PERMITTED_DATA_INCREASE, ProgramResult, SUCCESS},
     solana_program_test::{ProgramTest, processor},
     solana_pubkey::Pubkey,
     solana_rent::Rent,
     solana_signer::Signer,
     solana_system_interface::{instruction as system_instruction, program as system_program},
-    solana_sysvar::Sysvar,
     solana_transaction::Transaction,
     std::slice,
 };
@@ -93,7 +92,11 @@ fn invoke_create_account(
     let payer_info = next_account_info(account_info_iter)?;
     let create_account_info = next_account_info(account_info_iter)?;
     let system_program_info = next_account_info(account_info_iter)?;
-    let rent = Rent::get()?;
+    let mut rent = Rent::default();
+    assert_eq!(
+        solana_program_test::sol_get_rent_sysvar(&mut rent as *mut _ as *mut u8),
+        SUCCESS,
+    );
     let minimum_balance = rent.minimum_balance(MAX_PERMITTED_DATA_INCREASE);
     invoke(
         &system_instruction::create_account(