+++ runtime/src/stake_history.rs
@@ -1,10 +1,10 @@
 //! This module implements clone-on-write semantics for the SDK's `StakeHistory` to reduce
 //! unnecessary cloning of the underlying vector.
-pub use solana_stake_interface::stake_history::StakeHistoryGetEntry;
+pub use solana_stake_history::StakeHistoryGetEntry;
 use {
     serde::{Deserialize, Serialize},
     solana_clock::Epoch,
-    solana_stake_interface::stake_history::{self, StakeHistoryEntry},
+    solana_stake_history::StakeHistoryEntry,
     std::{
         ops::{Deref, DerefMut},
         sync::Arc,
@@ -31,7 +31,7 @@ impl DerefMut for StakeHistory {
 }
 
 /// The inner type, which is the SDK's stake history
-type StakeHistoryInner = stake_history::StakeHistory;
+type StakeHistoryInner = solana_stake_history::StakeHistory;
 
 impl StakeHistoryGetEntry for StakeHistory {
     fn get_entry(&self, epoch: Epoch) -> Option<StakeHistoryEntry> {
@@ -41,7 +41,7 @@ impl StakeHistoryGetEntry for StakeHistory {
 
 #[cfg(test)]
 mod tests {
-    use {super::*, solana_stake_interface::stake_history::StakeHistoryEntry};
+    use super::*;
 
     fn rand_stake_history_entry() -> StakeHistoryEntry {
         StakeHistoryEntry {